Добавил controllers и перенес всю отрисовку Board
